+++ AlarmClock/Views And Controllers/TheBestClockAboutScreenViewController.swift
@@ -87,7 +87,7 @@ class TheBestClockAboutScreenViewController: UIViewController {
                 if let versionTemp = plistDictionary["CFBundleShortVersionString"] as? NSString {
                     appVersion = versionTemp as String
                 }
-                if let versionTemp = plistDictionary["CFBundleDisplayName"] as? NSString {
+                if let versionTemp = plistDictionary["CFBundleName"] as? NSString {
                     appName = versionTemp as String
                 }
             }