+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Changed the prefs key, so we will get new prefs.
 - Changed the brightness sliders, so they have a larger hit test area.
 - The brightness sliders now extend further up.
+- Added a placeholder for German localization.
 
 **Version 1.0.0.2016** *(December 18, 2018)*
 - Fixed a minor cosmetic bug, where switching out of the app would not always restore the correct original brightness level.